python_utils/open_ai_dbt_doc_gen/pyproject.toml
[tool.poetry]
name = "open-ai-dbt-doc-gen"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "open_ai_dbt_doc_gen"}]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.27.6"
python-dotenv = "^1.0.0"
PyYAML = "^6.0"
ruamel-yaml = "^0.17.26"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

